Restructure stopper: inherit from base and implement stopping rule.
